requirements.txt
FastAPI
datetime
timedelta
uvicorn
requests
oneagent-sdk
setuptools
wheel
math